% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{transactions}
\alias{transactions}
\title{Transactions' information}
\format{
A data frame with 15379 rows and 5 variables
}
\usage{
data(transactions)
}
\description{
A dataset containing information about 15379 transactions of a certain bank.
}
\details{
\itemize{
\item nameOrig, character. ID of the client that initiated the transaction.
\item nameDest, character. ID of the client that received the transaction.
\item amount, numeric. Amount of money involved in the transaction in euros (€).
\item isFraud, boolean. Indicator whether the transaction was reported (1) or not (0).
\item transactionDate, character. Date of the transaction.
}
}
